Cleanup, util fn consolidation
