+++ lexicon/static/jsx/search/app.jsx
@@ -1,5 +1,3 @@
-// import Plyr from 'plyr';
-
 import React from 'react';
 import ReactDOM from 'react-dom';
