+++ mesolex/views.py
@@ -5,32 +5,11 @@ from django.shortcuts import render
 
 from lexicon.forms import formset_for_lg
 from narratives.forms import SoundMetadataQueryComposerFormset
-from mesolex.config import DEFAULT_LANGUAGE, LANGUAGES
+from mesolex.config import LANGUAGES
 from mesolex.utils import (
+    get_default_data_for_lg,
     ForceProxyEncoder,
 )
-
-def get_default_data_for_narratives():
-    language = LANGUAGES['narratives']
-    
-    return [{
-        'filter': 'begins_with',
-        'filter_on': language['filterable_fields'][0]['field'],
-        'operator': 'and',
-        'query_string': '',
-    }]
-
-
-def get_default_data_for_lg(language):
-    if language is None:
-        language = LANGUAGES[DEFAULT_LANGUAGE]
-    
-    return [{
-        'filter': 'begins_with',
-        'filter_on': language['filterable_fields'][0]['field'],
-        'operator': 'and',
-        'query_string': '',
-    }]
 
 
 def home(request, *args, **kwargs):
@@ -51,7 +30,7 @@ def home(request, *args, **kwargs):
         },
         'narratives': {
             'formset': SoundMetadataQueryComposerFormset(),
-            'formset_data': json.dumps(get_default_data_for_narratives()),
+            'formset_data': json.dumps(get_default_data_for_lg(LANGUAGES['narratives'])),
             'formset_global_filters_form_data': json.dumps({}),
             'formset_datasets_form_data': json.dumps({}),
             'formset_errors': json.dumps([]),
